pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "monitor-bot"
version = "1.0.0"
description = "Twitter and Solana monitoring bot with WeChat notifications"
readme = "README.md"
authors = [
    { name = "Monitor Bot Team" }
]
license = { text = "MIT" }
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "sqlalchemy>=2.0.23",
    "alembic>=1.12.1",
    "apscheduler>=3.10.4",
    "psycopg2-binary>=2.9.9",
    "requests>=2.31.0",
    "aiohttp>=3.9.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-multipart>=0.0.6",
    "loguru>=0.7.2",
    "solana>=0.30.2",
    "python-dotenv>=1.0.0",
    "base58>=2.1.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "httpx>=0.25.2",
    "black>=23.11.0",
    "isort>=5.12.0",
    "flake8>=6.1.0",
]

[project.urls]
Homepage = "https://github.com/your-org/monitor-bot"
Documentation = "https://github.com/your-org/monitor-bot#readme"
Repository = "https://github.com/your-org/monitor-bot.git"
Issues = "https://github.com/your-org/monitor-bot/issues"

[tool.black]
line-length = 100
target-version = ["py39"]

[tool.isort]
profile = "black"
line_length = 100

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
pythonpath = ["src"]

[dependency-groups]
dev = [
    "pytest-asyncio>=1.1.0",
    "pytest-mock>=3.14.1",
]
